Refactor report format handling by removing Enum definition from cli.py and consolidating it in formats.py
